resolve commits #PR68
