+++ app/src/Validators/regoster.ts
@@ -1,0 +1,21 @@
+import { z } from "zod";
+
+export const signupUserSchemaValidator = z.object({
+  userEmail: z.string().email({
+    message: "Invalid email",
+  }),
+  password: z.string().min(8, {
+    message: "Password must be atleast 8 characters long"
+  }).max(16, {
+    message: "Password must be atmost 16 characters long"
+  }),
+  userName: z.string().min(2, {
+    message: "Name should be atleast 2 characters long"
+  }),
+  contact: z.string().regex(/^\d{10}$/, {
+    message: "Contact number must be exactly 10 digits"
+  })
+
+});
+
+export type SignupUserRequest = z.infer<typeof signupUserSchemaValidator>;
