+++ app/src/app/api/feedback/route.ts
@@ -1,4 +1,5 @@
 import { validateRequest } from "@/auth";
+import { FEEDBACK_REQUEST_FAILED_MESSAGE } from "@/constants/messages";
 import { db } from "@/db";
 import { analysisFeedback, feedbackPayload } from "@/db/schema";
 import { z } from "zod";
