+++ app/src/components/DetailedTranscription.tsx
@@ -3,17 +3,13 @@
 import { format } from "date-fns";
 import { useCallback, useEffect, useRef, useState } from "react";
 import { Button } from "@/components/ui/button";
-import { TranscriptionsType } from "@/db/schema";
 import { PauseCircleIcon, PlayCircleIcon } from "lucide-react";
 import { Card } from "./ui/card";
 import { getAudioDuration } from "@/utils/recording";
 import Markdown from "react-markdown";
 import { cn } from "@/lib/utils";
 import { primaryFont } from "@/fonts";
-
-interface DetailedTranscriptionProps {
-  transcription: TranscriptionsType;
-}
+import { DetailedTranscriptionProps } from "@/types/transcriptions";
 
 const DetailedTranscription = ({
   transcription,
