+++ app/src/constants/app.ts
@@ -1,0 +1,9 @@
+import { Metadata } from "next";
+
+export const metadata: Metadata = {
+    title: "AiCruit",
+  };
+
+export const dynamic = "force-dynamic";
+export const fetchCache = "force-no-store";
+export const revalidate = 0;
