+++ app/src/types/app.ts
@@ -1,0 +1,5 @@
+import { User } from "lucia";
+
+export interface HeaderProps {
+  user: User | null;
+}
