+++ app/src/types/transcriptions.ts
@@ -1,3 +1,5 @@
+import { TranscriptionsType } from "@/db/schema";
+
 export type userTranscriptions = {
     id: string;
     documentName: string;
@@ -11,4 +13,32 @@ export type segment = {
   start: number;
   end: number;
   text: string;
+}
+
+export interface PageProps {
+  params: {
+    transcription_id: string;
+  };
+}
+
+export interface CustomBarChartProps {
+  data: {
+    labels: string[];
+    datasets: {
+      label: string;
+      data: number[];
+      borderColor: string[];
+      backgroundColor: string[];
+      fill: boolean;
+    }[];
+  };
+  maxY: number;
+}
+
+export interface DetailedTranscriptionProps {
+  transcription: TranscriptionsType;
+}
+
+export interface feedBackProps {
+  analysisId: string
 }
